+++ tecpg/__init__.py
@@ -1,1 +1,1 @@
-__version__ = '1.20.1-dev'
+__version__ = '1.21.0-dev'